bbox: size_t to more specific, rename

+++ src/bbox.cpp
@@ -1,41 +1,41 @@
-#include <cstdlib>
 #include "bbox.h"
+
+#include <stdint.h>
+#include <algorithm>
 
 
 using namespace bbox;
 
 // private:
-size_t BBox::s_img_width = 0;
+uint16_t BBox::s_img_width = 0;
 
 // public:
-BBox::BBox(const size_t idx, const size_t width, const size_t height) {
-    const size_t row = idx / s_img_width;
-    const size_t col = idx - row * s_img_width;
+BBox::BBox(const uint32_t idx, const uint16_t width, const uint16_t height) {
+    const uint16_t row = idx / s_img_width;
+    const uint16_t col = idx - row * s_img_width;
     topleft_X = col;
     topleft_Y = row;
     bottomright_X = col + width;
     bottomright_Y =  row + height;
 }
 
-void BBox::setImgWidth(const size_t img_width) {
+void BBox::setImgWidth(const uint16_t img_width) {
     s_img_width = img_width;
 }
 
- size_t BBox::width() const {
+ uint16_t BBox::width() const {
     return bottomright_X - topleft_X;
 }
 
- size_t BBox::height() const {
+ uint16_t BBox::height() const {
     return bottomright_Y - topleft_Y;
 }
 
-void BBox::merge(const size_t idx) {
-    size_t row = idx / s_img_width;
-    size_t col = idx - row * s_img_width;
-    // topleft_X = std::min(topleft_X, col);
-    // topleft_Y = std::min(topleft_Y, row);
-    bottomright_X = std::max(bottomright_X, col + 1);
-    bottomright_Y = std::max(bottomright_Y, row + 1);
+void BBox::merge(const uint32_t idx) {
+    uint16_t row = idx / s_img_width;
+    uint16_t col = idx - row * s_img_width;
+    bottomright_X = std::max(bottomright_X, (uint16_t)(col + 1));
+    bottomright_Y = std::max(bottomright_Y, (uint16_t)(row + 1));
 }
 
 void BBox::merge(const BBox other) {
